pyproject.toml
[tool.poetry]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
py-ulid = "^1.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.6"
alembic = "^1.14.0"
dependency-injector = "^4.44.0"
pydantic = {extras = ["email"], version = "^2.10.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
